Call fdatasync() after writing to the real file
If O_SYNC was not passed to jopen(), then we could loose data because
jtrans_commit() assumes writes to the file are synchronous. That behaviour
was undocumented, so the user could have assumed O_SYNC was not necessary.

This patch implements a much more sane semantic by calling fdatasync() on
the file fd after writing the transaction when lingering transactions are
not enabled, avoiding the need to open the file with O_SYNC.

+++ libjio/trans.c
@@ -416,6 +416,9 @@ ssize_t jtrans_commit(struct jtrans *ts)
 		ts->fs->ltrans = linger;
 		pthread_mutex_unlock(&(ts->fs->ltlock));
 	} else {
+		if (fdatasync(ts->fs->fd) != 0)
+			goto rollback_exit;
+
 		/* the transaction has been applied, so we cleanup and remove
 		 * it from the disk */
 		unlink(name);
